# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice System API
  description: |
    REST API for the invoice reconciliation platform (CSV/XLSX accounting
    imports, reconciliation runs, notification emails, audit trail, users,
    company and application settings).

    ## Versioning

    * **Stable, documented namespace:** `/api/v1` (this document).
    * **Compatibility namespace:** `/api` resolves to the *same* handlers
      and remains supported so existing clients (including the bundled SPA
      frontend) keep working unchanged. No business logic is duplicated;
      both prefixes are registered against identical view functions.

    ## Response contract

    Every endpoint returns JSON. Success payloads use:

    ```json
    { "success": true, "message": "...", "data": { } }
    ```

    Error payloads always use the stable error envelope:

    ```json
    {
      "success": false,
      "message": "Human readable reason",
      "status": 400,
      "code": "BAD_REQUEST"
    }
    ```

    See the `Error` schema and the `codes` values in the components section.

    ## Correlation

    Every response carries an `X-Request-Id` header. A client-supplied value
    (alphanumeric/dot/dash, up to 64 chars) is echoed back; otherwise a
    generated UUID is returned. Use it to correlate client requests with
    server logs.

    ## Deprecation

    Superseded endpoints (marked `deprecated: true`) keep working unchanged,
...
    major version.

    ## Money and dates

...
  version: "1.0.0"
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /reconciliation/runs:
    post:
      tags: [Reconciliation]
      operationId: startReconciliationRun
      summary: Start a reconciliation run for a company and period.
      description: |
        Idempotent: starting the same company + period that already has a run
        returns the existing run with HTTP 200 instead of 409 Conflict. A new
        run is created only when no run exists for the period, and returns 201.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period]
              properties:
                period:
                  type: string
                  pattern: "^(\\d{4}-\\d{2})$"
                  example: "2026-01"
                  description: Billing period (YYYY-MM).
                money_tolerance:
                  type: string
                  example: "0.50"
                  description: Optional monetary tolerance (decimal string).
      responses:
        "200":
          description: Existing run returned (idempotent replay).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "201":
          description: New run created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    get:
      tags: [Reconciliation]
      operationId: listReconciliationRuns
      summary: List reconciliation runs for the current company.
      description: Offset-based pagination.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Runs list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          runs:
                            type: array
                            items:
                              type: object
                              additionalProperties: true
                          count: { type: integer }
                          limit: { type: integer }
                          offset: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /reconciliation/runs/{run_id}:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: getReconciliationRun
      summary: Fetch a run and its per-status result counts.
      responses:
        "200":
          description: Run detail with counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reconciliation/runs/{run_id}/summary:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: getReconciliationSummary
      summary: Structured report summary for a run.
      responses:
        "200":
          description: Summary payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reconciliation/runs/{run_id}/results:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: getReconciliationResults
      summary: Invoice-level results for a run.
      description: |
        With no report query params a flat `{results: [...]}` list is
        returned (detail-page shape). With any of page/page_size/match_status/
        uuid/invoice_number/date_from/date_to present, a paginated envelope
        `{items, page, page_size, total, total_pages}` is returned instead.
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - in: query
          name: match_status
          schema: { type: string }
        - in: query
          name: uuid
          schema: { type: string }
        - in: query
          name: invoice_number
          schema: { type: string }
        - in: query
          name: date_from
          schema: { type: string, format: date }
        - in: query
          name: date_to
          schema: { type: string, format: date }
      responses:
        "200":
          description: Results payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reconciliation/runs/{run_id}/errors:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: getReconciliationErrors
      summary: Field/rule-level errors for a run.
      description: |
        Flat `{errors: [...]}` without report params; paginated envelope with
        page/page_size/error_type/source_type present.
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - in: query
          name: error_type
          schema: { type: string }
        - in: query
          name: source_type
          schema: { type: string, enum: [account, tax] }
      responses:
        "200":
          description: Errors payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reconciliation/runs/{run_id}/results/export:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: exportResults
      summary: Export a run's results as CSV or XLSX.
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        "200":
          description: Binary file download.
          content:
            text/csv:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reconciliation/runs/{run_id}/errors/export:
    parameters:
      - { $ref: "#/components/parameters/RunId" }
    get:
      tags: [Reconciliation]
      operationId: exportErrors
      summary: Export a run's errors as CSV or XLSX.
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        "200":
          description: Binary file download.
          content:
            text/csv:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token issued by POST /auth/login. Include as
        `Authorization: Bearer <token>`.
  parameters:
...
    RunId:
      in: path
      name: run_id
      required: true
      schema: { type: integer }
    Page:
      in: query
      name: page
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      in: query
      name: page_size
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
...
  schemas:
    SuccessBody:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
        message:
          type: string
          description: Optional human-readable text (not required for the contract).
        data:
          description: >-
            Contract payload field. Present on endpoints that return data;
            omitted only for signal-only actions (logout, delete, password
            change). When present it is always an object or array, never a raw
            scalar.
          nullable: true
    Error:
      type: object
      required: [success, message, status, code]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        status: { type: integer }
        code:
          type: string
          description: >-
            Stable machine-readable error code. `APP_ERROR` is a reserved
            internal fallback and must never be returned for the documented
            statuses below.
          enum:
            - BAD_REQUEST
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - TOKEN_REQUIRED
            - TOKEN_EXPIRED
            - INVALID_TOKEN
            - TOKEN_REVOKED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VALIDATION_ERROR
            - RATE_LIMITED
            - DATABASE_ERROR
            - INTERNAL_ERROR
            - BAD_GATEWAY
            - SERVICE_UNAVAILABLE
            - MISSING_FIELD
...
  responses:
...
    BadRequest:
      description: Invalid request.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Missing, expired, revoked or invalid token.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but role insufficient.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...